requirements.txt
nltk==3.2.1
numpy==1.11.0
pyenchant==1.6.6
scikit-learn==0.17.1
scipy==0.17.1
sklearn==0.0
